Working client.
Welp, I know it works. I don't have any tests. Some of this could be
tested, but a lot of it I'm not so sure about until I have a server
implementation to go along with it (and even then I'm unsure, since
it'll still need an actual oAuth provider...). One big gotcha is that
this uses a G5 fork of the oauth2 library because of reasons you can
read about there.