Use BrokenAuth setter they mercifully added to oauth2.
Previously there was an uneditable list of "broken" auth implementations
that don't adhere perfectly to the spec.  And no way to edit the list
because it's private to the package. Forking it and changing the
hardcoded list was the only way. They've relented and added this
RegisterBrokenAuthHeaderProvider function. Just in time, since I was
having awful conflicts from my fork not matching some libraries that
required tokens passed of the original version's type.