Stop hardcoding registration of "broken" auth endpoints.
You can read the comments for more info, but now I am hopefully just
registering whichever auth server you're using as broken, instead of
hardcoding a list. We are changing the dev-auth server's URL and it was
*really* hard to troubleshoot why it was broken. For grins (and to fix
a discrepency with how I implemented this vs. how the ruby one is
implemented), it can now take an endpoint name with or without a
protocol in it.